Adds git describe command.
The string returned by git describe is useful for versioning, but at least one commit needs to be tagged.

A file can be passed to GitEnv() and the repo containing that file will be used. This is useful to get the directory of a running program, even if you are not running from that directory.  For example `GitEnv(__file__)` will get the repo of the current running file.